Inital domain model scaffolded to the database

+++ Roz.Identity.EntityFramework/UserLogin.cs
@@ -2,5 +2,5 @@ using Microsoft.AspNet.Identity.EntityFramework;
 
 namespace Roz.Identity.EntityFramework
 {
-    public class UserLogin : IdentityUserLogin<long> { }
+    public class UserLogin : IdentityUserLogin<int> { }
 }
