+++ Roz.Identity.EntityFramework/UserStore.cs
@@ -3,7 +3,7 @@ using Microsoft.AspNet.Identity.EntityFramework;
 
 namespace Roz.Identity.EntityFramework
 {
-    public class UserStore : UserStore<User, Role, long, UserLogin, UserRole, UserClaim>
+    public class UserStore : UserStore<User, Role, int, UserLogin, UserRole, UserClaim>
     {
         public UserStore(DbContext context)
             : base(context)
