+++ Roz.WebApp/App_Start/Services/ApplicationSignInManager.cs
@@ -12,7 +12,7 @@ namespace Roz.WebApp.Services
     // Configure the application user manager used in this application. UserManager is defined in ASP.NET Identity and is used by the application.
 
     // Configure the application sign-in manager which is used in this application.
-    public class ApplicationSignInManager : SignInManager<User, long>
+    public class ApplicationSignInManager : SignInManager<User, int>
     {
         public ApplicationSignInManager(ApplicationUserManager userManager, IAuthenticationManager authenticationManager)
             : base(userManager, authenticationManager)
