+++ Roz.WebApp/Controllers/AccountController.cs
@@ -180,7 +180,7 @@ namespace Roz.WebApp.Controllers
         //
         // GET: /Account/ConfirmEmail
         [AllowAnonymous]
-        public async Task<ActionResult> ConfirmEmail(long userId, string code)
+        public async Task<ActionResult> ConfirmEmail(int userId, string code)
         {
             if (userId == default(long) || code == null)
             {
